+++ controllers/expose/ingressController/controller.go
@@ -3,44 +3,48 @@ package ingressController
 import (
 	"fmt"
 
-	"github.com/cloudogu/k8s-service-discovery/v2/controllers/expose/ingressController/nginx"
+	"github.com/cloudogu/k8s-service-discovery/v2/controllers/expose/ingressController/traefik"
 	ctrl "sigs.k8s.io/controller-runtime"
 )
 
 const (
-	DefaultIngressController = nginx.GatewayControllerName
+	DefaultIngressController = traefik.GatewayControllerName
 )
 
 type Dependencies struct {
-	Controller         string
-	ConfigMapInterface configMapInterface
-	IngressInterface   ingressInterface
-	IngressClassName   string
+	Controller       string
+	IngressInterface ingressInterface
+	IngressClassName string
+	TraefikInterface traefikInterface
+	Namespace        string
 }
 
 func ParseIngressController(deps Dependencies) IngressController {
 	switch deps.Controller {
-	case nginx.GatewayControllerName:
-		return nginx.NewNginxController(nginx.IngressControllerDependencies{
-			ConfigMapInterface: deps.ConfigMapInterface,
-			IngressInterface:   deps.IngressInterface,
-			IngressClassName:   deps.IngressClassName,
-			ControllerType:     nginx.GatewayControllerName,
+	case traefik.GatewayControllerName:
+		return traefik.NewTraefikController(traefik.IngressControllerDependencies{
+			IngressInterface: deps.IngressInterface,
+			IngressClassName: deps.IngressClassName,
+			ControllerType:   traefik.GatewayControllerName,
+			TraefikInterface: deps.TraefikInterface,
+			Namespace:        deps.Namespace,
 		})
-	case nginx.IngressControllerName:
-		return nginx.NewNginxController(nginx.IngressControllerDependencies{
-			ConfigMapInterface: deps.ConfigMapInterface,
-			IngressInterface:   deps.IngressInterface,
-			IngressClassName:   deps.IngressClassName,
-			ControllerType:     nginx.IngressControllerName,
+	case traefik.IngressControllerName:
+		return traefik.NewTraefikController(traefik.IngressControllerDependencies{
+			IngressInterface: deps.IngressInterface,
+			IngressClassName: deps.IngressClassName,
+			ControllerType:   traefik.IngressControllerName,
+			TraefikInterface: deps.TraefikInterface,
+			Namespace:        deps.Namespace,
 		})
 	default:
 		ctrl.Log.WithName("k8s-service-discovery.ParseIngressController").Error(fmt.Errorf("could not parse ingress controller %q. using default: %q", deps.Controller, DefaultIngressController), "unknown ingress controller")
-		return nginx.NewNginxController(nginx.IngressControllerDependencies{
-			ConfigMapInterface: deps.ConfigMapInterface,
-			IngressInterface:   deps.IngressInterface,
-			IngressClassName:   deps.IngressClassName,
-			ControllerType:     DefaultIngressController,
+		return traefik.NewTraefikController(traefik.IngressControllerDependencies{
+			IngressInterface: deps.IngressInterface,
+			IngressClassName: deps.IngressClassName,
+			ControllerType:   DefaultIngressController,
+			TraefikInterface: deps.TraefikInterface,
+			Namespace:        deps.Namespace,
 		})
 	}
 }