+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+> [!IMPORTANT]
+> Breaking change!
+> New compatible version of k8s-ces-gateway is required (3.x.x)
+
+### Changed
+- [#98] use traefik resources instead of nginx-ingress
 
 ## [v5.0.1] - 2026-02-27
 ### Changed